servers/bitwindow: rename binary to bitwindowd
