+++ servers/bitwindow/api/bitcoind/bitcoind.go
@@ -7,8 +7,8 @@ import (
 	"slices"
 
 	"connectrpc.com/connect"
-	pb "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/bitcoind/v1"
-	rpc "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/bitcoind/v1/bitcoindv1connect"
+	pb "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/bitcoind/v1"
+	rpc "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/bitcoind/v1/bitcoindv1connect"
 	corepb "github.com/barebitcoin/btc-buf/gen/bitcoin/bitcoind/v1alpha"
 	coreproxy "github.com/barebitcoin/btc-buf/server"
 	"github.com/btcsuite/btcd/btcutil"
