+++ servers/bitwindow/api/wallet/wallet.go
@@ -9,13 +9,13 @@ import (
 	"time"
 
 	"connectrpc.com/connect"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/bdk"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/drivechain"
-	validatorrpc "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/cusf/mainchain/v1/mainchainv1connect"
-	drivechainpb "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/drivechain/v1"
-	drivechainrpc "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/drivechain/v1/drivechainv1connect"
-	pb "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/wallet/v1"
-	rpc "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/wallet/v1/walletv1connect"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/bdk"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/drivechain"
+	validatorrpc "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/cusf/mainchain/v1/mainchainv1connect"
+	drivechainpb "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/drivechain/v1"
+	drivechainrpc "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/drivechain/v1/drivechainv1connect"
+	pb "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/wallet/v1"
+	rpc "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/wallet/v1/walletv1connect"
 	corepb "github.com/barebitcoin/btc-buf/gen/bitcoin/bitcoind/v1alpha"
 	coreproxy "github.com/barebitcoin/btc-buf/server"
 	"github.com/btcsuite/btcd/btcutil"
