+++ servers/bitwindow/gen/cusf/mainchain/v1/mainchainv1connect/wallet.connect.go
@@ -10,7 +10,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/cusf/mainchain/v1"
+	v1 "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/cusf/mainchain/v1"
 	http "net/http"
 	strings "strings"
 )
