+++ servers/bitwindow/gen/cusf/sidechain/v1/sidechainv1connect/sidechain.connect.go
@@ -10,7 +10,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/cusf/sidechain/v1"
+	v1 "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/cusf/sidechain/v1"
 	http "net/http"
 	strings "strings"
 )
