+++ servers/bitwindow/gen/wallet/v1/walletv1connect/wallet.connect.go
@@ -8,7 +8,7 @@ import (
 	connect "connectrpc.com/connect"
 	context "context"
 	errors "errors"
-	v1 "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/wallet/v1"
+	v1 "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/wallet/v1"
 	emptypb "google.golang.org/protobuf/types/known/emptypb"
 	http "net/http"
 	strings "strings"
