+++ servers/bitwindow/server/server.go
@@ -13,14 +13,14 @@ import (
 
 	"connectrpc.com/connect"
 	"connectrpc.com/grpcreflect"
-	api_bitcoind "github.com/LayerTwo-Labs/sidesail/drivechain-server/api/bitcoind"
-	api_drivechain "github.com/LayerTwo-Labs/sidesail/drivechain-server/api/drivechain"
-	api_wallet "github.com/LayerTwo-Labs/sidesail/drivechain-server/api/wallet"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/bdk"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/bitcoind/v1/bitcoindv1connect"
-	validatorrpc "github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/cusf/mainchain/v1/mainchainv1connect"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/drivechain/v1/drivechainv1connect"
-	"github.com/LayerTwo-Labs/sidesail/drivechain-server/gen/wallet/v1/walletv1connect"
+	api_bitcoind "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/api/bitcoind"
+	api_drivechain "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/api/drivechain"
+	api_wallet "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/api/wallet"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/bdk"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/bitcoind/v1/bitcoindv1connect"
+	validatorrpc "github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/cusf/mainchain/v1/mainchainv1connect"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/drivechain/v1/drivechainv1connect"
+	"github.com/LayerTwo-Labs/sidesail/servers/bitwindow/gen/wallet/v1/walletv1connect"
 	"github.com/barebitcoin/btc-buf/server"
 	"github.com/rs/zerolog"
 	"github.com/samber/lo"
